server_backup/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# OpenAI API
openai==1.12.0

# PDF Generation
reportlab==4.0.7

# Environment variables
python-dotenv==1.0.0

# Utilities
Pillow==10.2.0
psycopg2-binary==2.9.9
